% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmr_stratified.R
\name{cmr_stratified_from_rectangle}
\alias{cmr_stratified_from_rectangle}
\title{Stratified CMR from a variance rectangle}
\usage{
cmr_stratified_from_rectangle(
  rectangle,
  strata_share,
  control = list(),
  max_vertices = 65536L
)
}
\arguments{
\item{rectangle}{Stratified variance rectangle, a list with \code{lower} and
\code{upper} \verb{2 x S} matrices.}

\item{strata_share}{Named stratum population shares that sum to one.}

\item{control}{Optional list of solver controls for the general vertex
epigraph solver.}

\item{max_vertices}{Maximum number of hyperrectangle vertices to enumerate.}
}
\value{
A list of class \code{cmr_stratified} with assignment shares \code{pi}, \code{pi_matrix},
stratum sampling margins, within-stratum treatment margins, CMR certificate
\code{U_CMR}, checked rectangle, vertex diagnostics, and solver diagnostics.
}
\description{
Compute a stratified CMR allocation from a supplied cell-level variance
rectangle.
}
\examples{
strata_share <- c(A = 0.4, B = 0.6)
rect <- list(
  lower = rbind(treatment = c(A = 0.01, B = 0.04),
                control = c(A = 0.02, B = 0.03)),
  upper = rbind(treatment = c(A = 0.08, B = 0.12),
                control = c(A = 0.09, B = 0.10))
)
cmr_stratified_from_rectangle(rect, strata_share)

}
\seealso{
Other CMR rules: 
\code{\link{binary_rectangle_corners}()},
\code{\link{cmr_multiarm}()},
\code{\link{cmr_multiarm_from_rectangle}()},
\code{\link{cmr_multiple_outcomes}()},
\code{\link{cmr_proxy}()},
\code{\link{cmr_stratified}()},
\code{\link{cmr_two_arm}()},
\code{\link{cmr_unbounded}()},
\code{\link{cmr_unbounded_from_rectangle}()},
\code{\link{print.cmr_two_arm}()}

Other rectangle helpers: 
\code{\link{binary_rectangle_corners}()},
\code{\link{cmr_multiarm_from_rectangle}()},
\code{\link{cmr_unbounded_from_rectangle}()},
\code{\link{folded_binomial_pmf}()},
\code{\link{multiarm_variance_objective}()},
\code{\link{rectangle_bernoulli_binary}()},
\code{\link{rectangle_bounded_two_arm}()},
\code{\link{rectangle_multiarm}()},
\code{\link{rectangle_multiple_outcomes}()},
\code{\link{rectangle_proxy}()},
\code{\link{rectangle_stratified}()},
\code{\link{rectangle_unbounded}()},
\code{\link{stratified_variance_objective}()},
\code{\link{variance_bounds_bernoulli_exact}()},
\code{\link{variance_bounds_maurer_pontil}()},
\code{\link{variance_bounds_unbounded_mom}()}
}
\concept{CMR rules}
\concept{rectangle helpers}
