% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cmrdesign-package}
\alias{cmrdesign}
\alias{cmrdesign-package}
\title{cmrdesign: Conditional Minimax Regret design rules}
\description{
Applied Conditional Minimax Regret (CMR) design rules for pilot-informed
experiments. The package includes two-arm, shared-control multi-arm,
stratified, multiple-outcome, proxy/delayed-outcome, unbounded-outcome, and
pilot-planning workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://juancyamin.github.io/cmrdesign/}
  \item \url{https://github.com/juancyamin/cmrdesign}
  \item \url{https://arxiv.org/abs/2607.16982}
  \item Report bugs at \url{https://github.com/juancyamin/cmrdesign/issues}
}

}
\author{
\strong{Maintainer}: Juan C. Yamin \email{juan_yamin_silva@brown.edu} [copyright holder]

}
